training/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.66.0
wandb>=0.15.0
tensorboard>=2.14.0
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
python-dotenv>=1.0.0